+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# Release 17 (target)
+
+* Refactor Dock / Code cleanup
+* Add multiple-docks support
+* Add left, right, top layout
+* Add scrollwheel cycle/maximize/minimize
+
 # Release 16
 
 * Add preferred monitor